+++ README.md
@@ -139,7 +139,7 @@ You can view the [API reference for `FeatureDiscovery`](https://pub.dev/document
 
 ### `EnsureVisible`
 
-You can use the [`EnsureVisible` widget](https://pub.dev/documentation/feature_discovery/latest/feature_discovery/EnsureVisible-class.html) from the Flutter framework to automatically scroll to widgets that are inside of scrollable viewports when they are described during Feature Discovery:
+You can use the [`EnsureVisible` widget](https://pub.dev/documentation/feature_discovery/latest/feature_discovery/EnsureVisible-class.html) to automatically scroll to widgets that are inside of scrollable viewports when they are described during Feature Discovery:
 
 ```dart
 // You need to save an instance of a GlobalKey in order to call ensureVisible in onOpen.
