+++ lib/feature_discovery.dart
@@ -1,5 +1,4 @@
 library feature_discovery;
 
-//  export 'src/layout.dart';
 export 'src/feature_discovery.dart';
 export 'src/ensure_visible.dart';
